Structuring the pipeline
